LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/geolocation-API/enabled-on-self-origin-by-permissions-policy.https.sub-expected.txt

CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/resources/testdriver.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Feature policy 'Geolocation' check failed for element with origin 'https://127.0.0.1:9443' and allow attribute ''.

PASS Permissions-Policy header geolocation=(self) allows the top-level document.
PASS Permissions-Policy header geolocation=(self) allows same-origin iframes.
PASS Permissions-Policy header geolocation=(self) disallows cross-origin iframes.
